backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pydantic
requests
feedparser
python-dateutil
